# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ennote ennoworker API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /v1/skills/roots:
    get:
      operationId: listSkillRoots
      summary: List configured additional skills directories
      responses:
        "200":
          description: Configured roots ordered by priority
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [items]
                        properties:
                          items:
                            type: array
                            items: { $ref: "#/components/schemas/SkillRoot" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: createSkillRoot
      summary: Add an additional skills directory (by path or ecosystem preset)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string }
                path: { type: string, description: "Explicit absolute path; or leave empty and use agentKind" }
                agentKind: { type: string, enum: [pi, claude, codex, cursor, generic], description: "Preset resolving to the ecosystem skills dir under the worker home" }
                priority: { type: integer, minimum: 1 }
                enabled: { type: boolean }
      responses:
        "201":
          description: Root created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SkillRoot" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/skills/roots/{rootID}:
    parameters:
      - name: rootID
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateSkillRoot
      summary: Patch name, path, agent kind, priority, or enabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string }
                path: { type: string }
                agentKind: { type: string, enum: [pi, claude, codex, cursor, generic] }
                priority: { type: integer, minimum: 1 }
                enabled: { type: boolean }
      responses:
        "200":
          description: Updated root
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SkillRoot" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSkillRoot
      summary: Remove an additional skills directory
      responses:
        "204":
          description: Root removed
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }

  # ——— Skills management ———
  /v1/skills:
    get:
      operationId: listSkills
      summary: List the merged skill catalog with pi-ecosystem install annotations
      parameters:
        - name: projectID
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Catalog skills, diagnostics, and (for a project) trust status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SkillListResult" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /v1/skills/disabled/{relPath}:
    parameters:
      - name: relPath
        in: path
        required: true
        schema: { type: string, description: Skill rel path inside the user skills root; may contain slashes }
    patch:
      operationId: toggleSkillDisabled
      summary: Set or clear disable-model-invocation on a skill's SKILL.md
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [disabled]
              properties:
                disabled: { type: boolean }
      responses:
        "204":
          description: Toggle applied
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/skills/search:
    get:
      operationId: searchSkills
      summary: Search the skills.sh registry (with npx find fallback)
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 50 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [results]
                        properties:
                          results:
                            type: array
                            items: { $ref: "#/components/schemas/SkillSearchResult" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /v1/skills/install:
    post:
      operationId: installSkill
      summary: Install a skill via the skills.sh CLI (global, or project when trusted)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [package, scope]
              properties:
                package: { type: string }
                scope: { type: string, enum: [global, project] }
                projectId: { type: string }
      responses:
        "200":
          description: Install result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [success]
                        properties:
                          success: { type: boolean }
                          output: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /v1/skills/check:
    post:
      operationId: checkSkillUpdates
      summary: Compare recorded version hashes against the remote registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                projectId: { type: string }
                package: { type: string }
                scope: { type: string, enum: [global, project] }
      responses:
        "200":
          description: Per-install update states
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [updates]
                        properties:
                          updates:
                            type: array
                            items: { $ref: "#/components/schemas/SkillUpdateResult" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/skills/update:
    post:
      operationId: updateSkill
      summary: Re-install an existing skill to the latest version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [package, scope]
              properties:
                projectId: { type: string }
                package: { type: string }
                scope: { type: string, enum: [global, project] }
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: false
                        required: [success]
                        properties:
                          success: { type: boolean }
                          output: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /v1/skills/remove/{relPath}:
    parameters:
      - name: relPath
        in: path
        required: true
        schema: { type: string, description: Skill rel path inside the user skills root; may contain slashes }
    post:
      operationId: removeSkill
      summary: Remove a globally installed skill
      responses:
        "204":
          description: Skill removed
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
...
    SkillListResult:
      type: object
      additionalProperties: false
      required: [skills, diagnostics, projectResourcesLoaded]
      properties:
        skills:
          type: array
          items: { $ref: "#/components/schemas/AnnotatedSkill" }
        diagnostics:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              level: { type: string }
              message: { type: string }
              relPath: { type: string }
              source: { type: string }
        projectResourcesLoaded: { type: boolean }
    AnnotatedSkill:
      type: object
      additionalProperties: false
      required: [name, description, filePath, baseDir, disableModelInvocation, sourceInfo, skillId, relPath]
      properties:
        name: { type: string }
        description: { type: string }
        filePath: { type: string }
        baseDir: { type: string }
        disableModelInvocation: { type: boolean }
        sourceInfo:
          type: object
          additionalProperties: false
          properties:
            source: { type: string }
            scope: { type: string }
        skillId: { type: string }
        relPath: { type: string }
        install: { $ref: "#/components/schemas/SkillInstallInfo" }
    SkillInstallInfo:
      type: object
      additionalProperties: false
      required: [package, scope, source, sourceType, canCheckForUpdates]
      properties:
        package: { type: string }
        scope: { type: string, enum: [global, project] }
        source: { type: string }
        sourceType: { type: string }
        skillsShUrl: { type: string }
        skillPath: { type: string }
        ref: { type: string }
        versionHash: { type: string }
        canCheckForUpdates: { type: boolean }
    SkillSearchResult:
      type: object
      additionalProperties: false
      required: [package]
      properties:
        package: { type: string }
        installs: { type: string }
        url: { type: string }
    SkillRoot:
      type: object
      additionalProperties: false
      required: [id, name, path, agentKind, priority, enabled, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        agentKind: { type: string }
        priority: { type: integer }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SkillUpdateResult:
      type: object
      additionalProperties: false
      required: [package, scope, state]
      properties:
        package: { type: string }
        scope: { type: string, enum: [global, project] }
        state: { type: string, enum: [up-to-date, update-available, error, unsupported] }
        currentVersion: { type: string }
        latestVersion: { type: string }
        message: { type: string }
...